Refactor gogen by moving around functions/tests a bit
